Providers 모듈화 > ItemList 관심사 분리
