+++ src/components/ComplexForm.tsx
@@ -1,9 +1,9 @@
-import { useState } from "react";
+import { memo, useState } from "react";
 import { renderLog } from "../utils";
 import { useNotification } from "../contexts";
 
 // ComplexForm 컴포넌트
-export const ComplexForm: React.FC = () => {
+export const ComplexForm: React.FC = memo(() => {
   renderLog("ComplexForm rendered");
   const { addNotification } = useNotification();
   const [formData, setFormData] = useState({
@@ -85,4 +85,4 @@ export const ComplexForm: React.FC = () => {
       </form>
     </div>
   );
-};
+});
