+++ src/contexts/ThemeContext.tsx
@@ -1,23 +1,11 @@
-import { createContext, ReactNode, useContext, useState } from "react";
+import { createContext, useContext } from "react";
 
 type ThemeContextProps = {
   theme: string;
   toggleTheme: () => void;
-}
-
-const ThemeContext = createContext<ThemeContextProps | null>(null);
-
-export const ThemeProvider = ({ children }: { children: ReactNode }) => {
-  const [theme, setTheme] = useState("light");
-
-  const toggleTheme = () => {
-    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
-  };
-
-  const value = { theme, toggleTheme };
-
-  return <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>
 };
+
+export const ThemeContext = createContext<ThemeContextProps | null>(null);
 
 export const useTheme = () => {
   const context = useContext(ThemeContext);
