+++ src/providers/AppProviders.tsx
@@ -1,0 +1,16 @@
+import React, { ReactNode } from "react";
+import { ThemeProvider } from "./ThemeProvider";
+import { NotificationProvider } from "./NotificationProvider";
+import { UserProvider } from "./UserProvider";
+
+export const AppProviders: React.FC<{ children: ReactNode }> = ({
+  children,
+}) => {
+  return (
+    <ThemeProvider>
+      <NotificationProvider>
+        <UserProvider>{children}</UserProvider>
+      </NotificationProvider>
+    </ThemeProvider>
+  );
+};
