+++ src/providers/NotificationProvider.tsx
@@ -1,0 +1,35 @@
+import { ReactNode, useState } from "react";
+import { NotificationContext } from "../contexts";
+
+type Notification = {
+  id: number;
+  message: string;
+  type: "info" | "success" | "warning" | "error";
+};
+
+export const NotificationProvider = ({ children }: { children: ReactNode }) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = (message: string, type: Notification["type"]) => {
+    const newNotification: Notification = {
+      id: Date.now(),
+      message,
+      type,
+    };
+    setNotifications((prev) => [...prev, newNotification]);
+  };
+
+  const removeNotification = (id: number) => {
+    setNotifications((prev) =>
+      prev.filter((notification) => notification.id !== id),
+    );
+  };
+
+  const value = { notifications, addNotification, removeNotification };
+
+  return (
+    <NotificationContext.Provider value={value}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
